Maintain unplaced_items list when adding/removing groups
The unplaced_progress_items list contains group names for items in
PROGRESS_ITEM_GROUPS, and sometimes needs to be expanded/packed when
these lists are modified to maintain the invariant that they contain the
group name if and only if all the items in the group are in the list

The add_ and remove_owned_item_or_item_group functions did not maintain
the unplaced_items lists when adding or removing entire groups. This
wasn't previously a problem since these functions were only used in a
temporary context for adding/removing groups that would be rolled back
after. This is preparatory work for the starting item randomizer which
uses these functions, and needs them to also maintain the invariant

It adds 2 helpers to do this maintenance by merging or splitting items
representing groups in lists of items (`expand_/pack_item_groups`).
A few existing functions that were doing the same thing have been
updated to use these helpers