Require completing the puzzle to enter DRC
In vanilla wind waker, you need an empty bottle to enter DRC, to water
some bomb flowers and blow up a boulder.
To experiment with not having DRC always sphere 0, we recreate a similar
puzzle by making the 3 bomb flowers after the pond before the entrance
to DRC withered, so that you need a way to go over the lava pit with the
stone heads that isn't sphere 0.
Watering the flowers is an option, but you can also bring your own
bombs, or a leaf to fly over.
We also remove the tweak that disables Din's pearl effect to offer
another option for going through