Unconditionally filter out non-progression sunken treasure
In the actual randomizer (not the UI), we always know whether sunken
treasure is progression or not once the chart randomizer has run.
Update `filter_locations_for_progression` to unconditionally exclude
non-progression sunken treasure, which fixes incorrect values for item
usefulness fractions in `get_items_by_usefulness_fraction` and
potentially other places. The static version (used in the UI) is
unchanged

There is no reason for this function to return locations that are known
to not be in logic. `filter_sunken_treasure` is kept as an override so
that it's possible to check for non-chart progression, but can now only
remove additional locations, instead of adding non-progression locations
to the list of progression locations when False

This fixes various issues where the Grappling Hook would always have a
maximum usefulness fraction of 2, because the item randomizer couldn't
filter with the correct granularity and thus always considered sunken
treasure location progression